backend/api/views.py
from api.filters import PropertyFilter
from api.models import Property
from api.serializers import PropertySerializer
from django_filters.rest_framework import DjangoFilterBackend
from rest_framework import filters, permissions, viewsets


class PropertyViewSet(viewsets.ModelViewSet):
    queryset = Property.objects.all()
    serializer_class = PropertySerializer
    permission_classes = [permissions.IsAuthenticated]
    filter_backends = [filters.SearchFilter, DjangoFilterBackend]
    search_fields = ["@full_address", "@class_description"]
    filterset_class = PropertyFilter
