backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Karina Boikova <karinaboikova@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
psycopg2 = "^2.9.9"
pandas = "^2.2.2"
openpyxl = "^3.1.5"
djangorestframework = "^3.15.2"
markdown = "^3.7"
django-filter = "^24.3"
pytest = "^8.3.2"
pytest-django = "^4.8.0"
factory-boy = "^3.3.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.6.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

target-version = "py312"

[tool.black]
target-version = ["py312"]

[tool.ruff]
# Enable Pyflakes `E` and `F` codes by default.
select = ["E", "F"]
ignore = ["E501", "E741"]
